Replace FetchContent with vcpkg for most things
Moved the codebase over to using vcpkg to handle nearly all dependencies.
FastNoise and vk-bootstrap both have custom port files, the former because
it isn't in vcpkg and the latter because I want to be able to use custom
versions.
The only dependency still using FetchContent is vuk because it seemed to
have non trivial cmake code and I can't be bothered to figure it out all in
this first go. Thats a problem for future me.