Attempted to get vuk through vcpkg
It configures successfully but fails to build, citing a missing dependency
on spirv-cross-core.lib (for windows)
I am not sure why main.exe is trying to link in one of vuk's dependencies
but that is a job for another day.

Note that to get vuk using vcpkg, a custom port of VMA was created which
referenced the fork vuk uses. This also required a patch file to make it rename the
vk_mem_alloc.h header since the main project uses VMA, although a newer build.

A very hacky mess, but it is a good learning experience on how packaging & vcpkg
all works together.