Successfully compiles on windows!
This commit makes the necessary changes to VUK and the project's cmake code
to make it compile & run again!

The issue was that for whatever reason, the dependencies that VUK needed were
not being provided. By making the OragneEngine also look for them (using
find_package) the compiler can link all the necessary dependencies. This does
require defining VMA_IMPLEMENTATION in main.cpp so VUK has definitions for the
vulkan memory allocator functions.

Also took a minute to remove commented out and unnecessary changes in the patch
file.